Remove moved code from naming.py
